[Untested] IR Sensor Voltage Reading
Added IR Sensor to RobotMap, added the getVoltage() fxn for the IR
Sensor to IntakeSubsystem as well as additions to SmartDashboard that
should show us the IR voltage, added a part to IntakeCommand that when
the difference between the current voltage and previous voltage goes
over a certain amount, then the intake will continue for .5 sec and then
stop. We need to choose a wait time, and choose a threshold value, as
well as just testing to make sure we understand voltage values, and
that's about it.