Portcullis/Cheval and Low Bar Autonomous Programs
Created 2 autonomous command groups: Portcullis (works for port and
cheval) and Low Bar (works for low bar). TODO: tune. The values for
distance and time and everything haven't been put in, mostly because we
haven't had much of a chance to test the robot with the proper field
setup, so WORK ON THAT. Low Bar Auto has not been tested yet