Integrated Pixy Auto
pixy code has been put into PixyAutoCommand. It lines up to the goal
pretty accurately, but we cannot be sure how accurately because due to
the fact that we have no carpets, there is sliding going on. Still, we
are able to traverse a defense (drive forward) and then aim from almost
any position to where we can shoot